laplace: Avoid returning infinite results
The internal computation of a Laplace transform might yield an infinite
result when done through manual integration, which probably means that
the integral does not converge. Returning this infinite value is
generally unhelpful.

Modify `_laplace_transform` to detect if the computed result contains
S.Infinity, S.NegativeInfinity, or S.ComplexInfinity. If so, return the
 unevaluated `LaplaceTransform(f, t, s)` object.

Fixes the failed test in laplace transform that is caused by the power
simplification rule.